login finished

+++ core/src/main/java/com/tw/controller/UsrController.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
+import javax.servlet.http.HttpSession;
+
 
 /**
  * Created by twer on 7/13/15.
@@ -17,22 +19,44 @@ public class UsrController {
     @Autowired
     private UserService userService;
 
+
     @RequestMapping(method = RequestMethod.GET)
-    public ModelAndView getUsers() {
-        return new ModelAndView("index", "usrs", userService.get_users());
+    public ModelAndView getUsers(HttpSession session) {
+
+        String loginStatement = (String) session.getAttribute("loginStatement");
+        if(loginStatement == "login"){
+            return new ModelAndView("index", "usrs", userService.get_users());
+        }else{
+            return new ModelAndView("redirect:/login");
+
+        }
     }
 
     @RequestMapping(method = RequestMethod.POST)
-    public ModelAndView insertUser(@RequestParam String name, String gender, int age, String email) {
-        Usr usr = new Usr(name, gender, age, email);
-        userService.insert_users(usr);
-        return new ModelAndView("redirect:/");
+    public ModelAndView insertUser(@RequestParam String name, String gender, int age, String email,HttpSession session) {
+        System.out.println("++++++++++++++");
+        String loginStatement = (String) session.getAttribute("loginStatement");
+        if(loginStatement == "login"){
+            Usr usr = new Usr(name, gender, age, email);
+            userService.insert_users(usr);
+            return new ModelAndView("redirect:/");
+        }else{
+            return new ModelAndView("redirect:/login");
+
+        }
     }
 
     @RequestMapping(value="/deleteUsr",method = RequestMethod.GET)
-    public ModelAndView deleteUsers(@RequestParam int id) {
-        userService.delete_user(id);
-        return new ModelAndView("redirect:/");
+    public ModelAndView deleteUsers(@RequestParam int id,HttpSession session) {
+        String loginStatement = (String) session.getAttribute("loginStatement");
+        if(loginStatement == "login"){
+            userService.delete_user(id);
+            return new ModelAndView("redirect:/");
+        }else{
+            return new ModelAndView("redirect:/login");
+
+        }
+
     }
 
     @RequestMapping(value="/updateUsr",method = RequestMethod.GET)
@@ -45,9 +69,17 @@ public class UsrController {
     }
 
     @RequestMapping(value="/updateUsr",method = RequestMethod.POST)
-    public ModelAndView updateUser(@RequestParam int id, String name, String gender, int age, String email) {
-        Usr usr = new Usr(id,name,gender,age,email);
-        userService.update_user(usr);
-        return new ModelAndView("redirect:/");
+    public ModelAndView updateUser(@RequestParam int id, String name, String gender, int age, String email,HttpSession session) {
+        String loginStatement = (String) session.getAttribute("loginStatement");
+        if(loginStatement == "login"){
+            Usr usr = new Usr(id,name,gender,age,email);
+            userService.update_user(usr);
+            return new ModelAndView("redirect:/");
+        }else{
+            return new ModelAndView("redirect:/login");
+
+        }
+
     }
+
 }
